Added mega and giga tiers of storage cells (TODO finish fluid)

+++ src/main/java/appeng/api/implementations/items/IStorageComponent.java
@@ -39,7 +39,7 @@ public interface IStorageComponent {
      *
      * @return number of bytes
      */
-    int getBytes(ItemStack is);
+    long getBytes(ItemStack is);
 
     /**
      * Just true or false for the item stack.
