+++ src/main/java/appeng/items/tools/powered/ColorApplicatorItem.java
@@ -423,12 +423,12 @@ public class ColorApplicatorItem extends AEBaseItem
     }
 
     @Override
-    public int getBytes(ItemStack cellItem) {
+    public long getBytes(ItemStack cellItem) {
         return 512;
     }
 
     @Override
-    public int getBytesPerType(ItemStack cellItem) {
+    public long getBytesPerType(ItemStack cellItem) {
         return 8;
     }
